third_party/blink/web_tests/editing/selection/internal-caret-rect-expected.txt

Bug 86390: Expose FrameSelection::absoluteCaretBounds via window.internals

This test demonstrates the usage of internals.absoluteCaretBounds()

abcd
PASS caretRects[0].left is 8
PASS caretRects[0].top is 160
PASS caretRects[0].width is 1
PASS caretRects[0].height is 20
PASS caretRects[1].left is 28
PASS caretRects[1].top is 160
PASS caretRects[1].width is 1
PASS caretRects[1].height is 20
PASS caretRects[2].left is 48
PASS caretRects[2].top is 160
PASS caretRects[2].width is 1
PASS caretRects[2].height is 20
PASS caretRects[3].left is 68
PASS caretRects[3].top is 160
PASS caretRects[3].width is 1
PASS caretRects[3].height is 20
PASS caretRects[4].left is 88
PASS caretRects[4].top is 160
PASS caretRects[4].width is 1
PASS caretRects[4].height is 20
PASS successfullyParsed is true

TEST COMPLETE
